POC: hard-code HMPO Webchat page / remove Content Item dependency
I think by trying to fit into the "content item rendering" model,
we're making things more complicated than they need to be. I argue
that we should decouple the Webchat page from Content Store
entirely, for the following reasons:

1. We have no means of publishing new 'Contact' content items now
   that Contacts Admin is being retired.
2. The concept of a 'Contact' document / schema type is therefore
   deprecated.
3. Hard-coding a copy of the old content item in government-frontend
   isn't great either, since the Content API could change in future
   and we'd then have to remember to update this item (it doesn't
   get updates for free).
4. There is only one 'contact' remaining - so maintaining the
   existing Contact controller/renderer is an unnecessary abstraction.
5. We already have a bunch of hard-coded Webchat config in the
   webchat.yaml file. We may as well add the two bits that came from
   Content Store (title and 'more info' / body), especially as the
   page has only ever existed in English and has no non-English
   audience.

This commit therefore:

- Deletes all the legacy Contact stuff that it can. (Be aware that
  `ContentItem::ContactDetails` is still needed, however, for
  WorldwideOrganisation contacts).
- Creates a minimal WebchatController and WebchatPresenter.
- Moves the shared _webchat.html.erb partial into its own top level
  webchat.html.erb template, called by the controller.

NB some of the tests will need updating as they're currently
dependent on `setup_and_visit_content_item("contact")`. We can
likely simplify these a lot.

Trello: https://trello.com/c/QrArPjwt/3675-move-hm-passport-office-out-of-contacts-admin
