+++ app/views/content_items/webchat.html.erb
@@ -1,0 +1,58 @@
+<% content_for :simple_header, true %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <%= render 'govuk_publishing_components/components/heading', @content_item.heading_and_context %>
+  </div>
+</div>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <div class="content-bottom-margin">
+      <%= render 'govuk_publishing_components/components/govspeak' do %>
+        <h2 id="webchat-title"><%= t("contact.webchat") %></h2>
+
+        <%= @content_item.webchat_body %>
+
+        <p>
+          <span class="js-webchat"
+            data-availability-url="<%= @content_item.webchat.availability_url %>"
+            data-open-url="<%= @content_item.webchat.open_url %>"
+            data-redirect="<%=  @content_item.webchat.open_url_redirect.present? ? 'true': 'false' %>"
+          >
+            <% if webchat_unavailable? %>
+              <%= unavailability_message %>
+            <% else %>
+              <span class="js-webchat-advisers-error">
+                <%= t("shared.webchat.technical_problem") %>
+              </span>
+              <span class="js-webchat-advisers-unavailable govuk-!-display-none">
+                <%= t("shared.webchat.closed") %>
+              </span>
+              <span class="js-webchat-advisers-busy govuk-!-display-none">
+                <%= t("shared.webchat.busy") %>
+              </span>
+              <span class="js-webchat-advisers-available govuk-!-display-none">
+                <%= t("shared.webchat.available") %>
+                <a
+                  href="#"
+                  data-redirect="<%=  @content_item.webchat.open_url_redirect.present? ? 'true' : 'false' %>"
+                  rel="external"
+                  class="js-webchat-open-button"
+                  data-module="ga4-link-tracker"
+                  data-ga4-link="<%= { event_name: "navigation", type: "webchat", text: t("shared.webchat.speak_to_adviser", locale: :en) }.to_json %>">
+                    <%= t("shared.webchat.speak_to_adviser") %>
+                </a>.
+              </span>
+            <% end %>
+          </span>
+        </p>
+        <% # This is inline in the source however slimmer will optimize this. %>
+        <%= javascript_include_tag "webchat", integrity: false, type: "module" %>
+      <% end %>
+    </div>
+  </div>
+  <%= render 'shared/sidebar_navigation' %>
+</div>
+
+<%= render 'shared/footer_navigation' %>
