+++ lib/webchat.yaml
@@ -3,3 +3,5 @@
   availability_url: https://d1y02qp19gjy8q.cloudfront.net/availability/18555309
   csp_connect_src: https://d1y02qp19gjy8q.cloudfront.net
   open_url_redirect: false
+  title: HM Passport Office webchat
+  more_info_webchat: <p>Monday to Friday, 8am to 8pm <br>\nWeekends and public holidays, 9am to 5:30pm</p>\n\n<p>Advisers may be busy at peak times.</p>\n
